Move CCT API mocks to their own file
